refactor/ change Messenger class to Module
