+++ src/types/Message.ts
@@ -1,0 +1,6 @@
+type Message = {
+  type: string;
+  payload?: any;
+};
+
+export default Message;
